Add Output.Filters.
This will store a list of filters to be performed on an output before committing a save to data storage (and thus before determining hashes to be checked against the conjecture).  No implementation yet, but this will describe i.e. if file mod times will be sanitized before commit (which will usually default to "god, yes").

Signed-off-by: Eric Myhre <hash@exultant.us>