Start tree walk library.
I need pre- and post-order visitors to walk a tree, and I need it for both the filesystem (stdlib walk only does pre-order traversal) and for our hash buckets.

It's almost like.... it's almost like I need generics to solve this correctly.  Fancy that.  That NEVER happens in golang.

Oh, wait, it's been a week since I last ran into a problem like this.  Yep, it was about due time for it to happen again.

Signed-off-by: Eric Myhre <hash@exultant.us>