Begin work on consistent serialization.
Metadata in fshash absolutely must serialize the same way, every time, in order for the hashing to make any sense.  We're going to use cbor because it's a sane and dense encoding, and can be coerced to be sufficiently deterministic in output... however, in order to obtain this, we're going to have to reach into the codec library we're using pretty deeply, because it doesn't do map orders deterministically by itself.

This is just a proof of concept and a quick test to make sure that as we add the real thing, we don't make mistakes in the serialization code such that we end up with something undecodable.  In practice, we wouldn't care about that or notice for now because we're just shovelling it into a hash rather than keeping it, but it's better to be consistently right nonetheless.

Signed-off-by: Eric Myhre <hash@exultant.us>