Remove some progress notes.
I was looking way too directly at bytes for a while there.  But it was fun!  Cbor is nice... packing "there is a string of length 3 coming up" into '0x63' is just fantastically precise and efficient when you multiply it out across all the short map keys in a (schemaless) message.

Signed-off-by: Eric Myhre <hash@exultant.us>