Adding uid+gid support and forced uid to dir tests
Make the fshash copying procedures apply uid and gid bits.

Regretably, this marks the line where we have to start requiring root priviledges in order to even run the tests.  So, this commit also contains test helpers for skipping if not uid=0, travis reconfigured to run tests as root, and arg passthru for goconvey (in case you want to run two instances, one for root tests).

Later in the game, much of this subject will be addressed by filters, but: that will still require root; I'd rather not break ground on adding them in this already-large PR; and hardcoding the uid settings (like already done with the modtiems) seems appropriate for these unit tests basic input tests since this area should be testable without requiring covering filters.

Signed-off-by: Eric Myhre <hash@exultant.us>