Run without a tty.
Our enclosing environment is not promised to be a tty (in fact, in CI itself right now, it's not (well, it... travis sorta has one.  But we still get an "inappropriate ioctl for device" complaint, which this fixes)).  And anyway, we probably won't want this eventually, since we'll be building our own logged hookups.

Signed-off-by: Eric Myhre <hash@exultant.us>