Add basic command output access.
This is super, super placeholdery, but enough to let us write a slightly more legitimate test that things are working in an executor (and stop ugly noise from bubbling out of tests).

This *doesn't* do buffering correctly, support mulitple readers, dtrt if you start reading before the job ends, proxy EOFs, or 100 other things.  Better systems for describing io streams from a task are going to be their own PR in scope, and the interface on def.Job will get much more expressive in the future.

Signed-off-by: Eric Myhre <hash@exultant.us>