Wow, can we maybe not quietly discard major errors about kernel features not supported?
I don't expect to hit this, but I also don't want to be completely in the dark if it does come up somewhere.

Indeed, this only increases the parity between the link-aware and link-transparent versions of the calls, since the link-transparent ones are already reporting ENOSYS errors, and in fact would have no defensible reason to exist separately from `os.Chtimes` if they didn't.

Signed-off-by: Eric Myhre <hash@exultant.us>