Start using fs.PlaceFile to replace some of the fshash logic.
Can't quite trivially replace the file case itself because the hasher stream is operating in the opposite direction to how tar readers tend to... ヽ(´ー｀)ノ

Also consistently look at the srcpath in fs_walk.  This mostly didn't make a difference since we're using the already-loaded fileinfo, but pointing at destpath there is just... odd.  And fix a duplicate/unnecessary readlink call.

Signed-off-by: Eric Myhre <hash@exultant.us>