Scan the rest of the interesting attributes.
This lets us remove a huge chunk of todos from fs_walk; the only conditional branch in behavior there now is whether or not it's dealing with a file, since that needs content hashing -- this is looking much better.  Within another iteration or so, they 'fspatch' package import should disappear from that file entirely.

Note that we are *not* keeping all-platform compatability here.  I'm sorry.  We'll correct this when we upgrade the build systems to at least build for a complete list of supported platforms; I have no intention of trying to roughly maintain allplat by eyeballing it without that piece of infrastructure.

Signed-off-by: Eric Myhre <hash@exultant.us>