Add basic hierarchy to output errors.
This is very similar to the prototype in the chroot executor and seems to be working out well.  Not sure if we'll change the output interface method sig to enforce spacemonkey errors, but maybe.  Table-driven tests making sure no output ever emits errors outside the appropriate heirarchy also seem like a reasonable future step.

Signed-off-by: Eric Myhre <hash@exultant.us>