Start documenting PlaceFile more thoroughly... and realize there's a whole category of issues around symlinks that are insufficiently addressed.
Basically any unpacking operation needs to be careful not to follow other symlinks.  A check for that is now added, but it's woefully inefficient due to the fact it's working on the scope of single files, whereas an overall process could do much fewer redundant stat calls, so... though it will be at the cost of PlaceFile being individually unsafe to use, this is probably not going to stay quite here.

Signed-off-by: Eric Myhre <hash@exultant.us>