Complete hashing tar output implementation.
First tests.  Starting to gun for table driven tests.  Currently this has some sketchy hacks with relative paths that the "reusable" tests and some of their inputs have to agree on... a factory pattern of some kind might make more sense here; gonna let that idea steep for a while.

Signed-off-by: Eric Myhre <hash@exultant.us>