Fix some of the sketchy hacks with test paths.
Need this OutputFactory indirection for other reasons as well: a lot of tests are going to use more than one output.  Currently, the tar outputs are basically stateless and in effect reusable (given the same paths are fine, anyway), but this is *not* going to be a valid assumption for all output implementations, so this indirection is necessary for getting a new output per use.

Signed-off-by: Eric Myhre <hash@exultant.us>