Use structural type for row.
We could squeeze a tiny bit more efficiency out of just having an int->something map, which would be almost always a byte slice but sometimes an out-of-band signal of another type.  But attempting that produced serialize and deserialize code that was looking suuuper ugly, and I doubt anyone else would ever be happy writing a parser for it.

This cborMuxRow struct might be removed from the hot path again in the long run in favor of hand rolled serializers for efficiency, but if nothing else, we'll probably keep it around for testing assertions.

Removed all the debug writers and stuff.

Signed-off-by: Eric Myhre <hash@exultant.us>