Fix returns.
I almost wonder if it would be useful to create a pattern of workers with report channels like this and give them an abort(err) method that sends the error report and then has a different kind of panic to quietly terminate the worker goroutine.  Of course, it would want generics again, so I'm going to sit on my hands instead.

Signed-off-by: Eric Myhre <hash@exultant.us>