Review and fix error handling consistency.
All `Output` implementations now use the same `try` pattern and the same filtering.  Previous direct usage of the done channel was making it easy to slip up and send irregular errors, so that is now shunned, and in correcting it new (actually meaningful!) error types showed up.

There are still errors that pass through the "unknown" path of the filter.  I like the idea of an error clearly indicating whether it was the scanning part or the output committing part that had an error, and probably want to drive that further in the future, but wanna let that idea stew for awhile before implementing it everywhere.  For now, at least everything hits the "unknown" path and still get categorized, as intended.

Signed-off-by: Eric Myhre <hash@exultant.us>