Update tar exec'ing implementations.
Use of gosh means their error handling is now consistent in all the ways gosh addressess (exit codes are accounted for, signals are accounted for, ptrace is accounted for, etc), and also uses a feature that means output is reported attached to an error in the case of the command reporting nonsuccess.

Direct connection of the tar commands to the repeatr process's stdin/out/err is curtailed.

Signed-off-by: Eric Myhre <hash@exultant.us>