Upgrade chroot executor to use gosh.
This lets us remove a bunch of duplicated exit detection code, and gets some other fixes in from gosh for things like better command-not-found edge case detection than repeatr previously had.

I did *not* change the nsinit executor because I'm not actually sure what to do with it.  I'm not sure nsinit proxies exit codes (and if so, it's automatically not lossless because nsinit errors gotta go somewhere), and similarly command-not-found for the job definition is a totally different thing than command-not-found for nsinit when setting up, and though we're not handling either of those now, I a lil wary of commiting new code that's still muddled on those.

Signed-off-by: Eric Myhre <hash@exultant.us>