Fix spin when waiting for more data in a stream.
This is not *at all* a clever implementation, and is in fact still spinning; but spinning once a millisecond still means a core isn't pegged.

A better implementation could use an in-memory lock if we decide this doesn't need to be functional across processes (but I suspect we'll want that later), or some fnotify stuff (except that needs fallbacks for when you can't set any more watches, so it would fall back to this mode *anyway*...)... it's a tricky subject; this should be "good enough" for now.

n.b. I can't really think of any ways to write a test for this :I

Signed-off-by: Eric Myhre <hash@exultant.us>