Add rootedness check to fshash.Bucket.
Some input formats from other systems will not typically include the root dir's properties.  For these systems, we'll just put in a sane default value.

It's not super clear if it would be better to do this as a sanitize step in the bucket implementations, or force each input implementation to do it individually; right now it's in the bucket code because it fit there, but it's not the prettiest (or most efficient) thing.  Open to change in the future.

Signed-off-by: Eric Myhre <hash@exultant.us>