Set a timeout for tar integration tests.
Slinging around full OS images does take a moment.  Tests are also *ridiculously* unoptimized and are scanning a whole OS into memory *twice* even *after* the real work, etc.

This is crazy too high and we should probably probably do some judicious skips for normal iterations.  In addition to getting test coverage without, you know, hucking all of ubuntu at it blindly.  Todo!

Although also note... cpu profiling seems to indicate that the single largest time sink here is the huffman encoding shuffling in the gzip extraction.  So.  That's... hrm.  Also archive/tar checks the crc32's per file, which consumes *the same amount of time as the sha512* that we're running.  I don't think there's a switch for it without forking the tar library, but it's like 10% of time; it's irritating enough either that or attempting to PR upstream might be worth it.

Signed-off-by: Eric Myhre <hash@exultant.us>