Set up job results precisely once; mutate.
Setting up the structure once outside of the error handling blocks and mutating references to it makes it much less easy to accidentally return a differently initialized structure and just generally averts weirdness.

Zero values also overlap with valid values (namely, with exit codes), so initializing an unambiguous value here is a big deal for maintaining appearances of sanity in error handling paths.  (As a specific example, assertions that the exit code is zero a.k.a. success should fail if the process never launched.)

n.b. I don't think all of these functions should be exported... but not making any changes right now.

Signed-off-by: Eric Myhre <hash@exultant.us>