Canonicalize paths to have dot-slash prefixes and trailing slash suffixes on directories.
Docstrings in the commit tell the story of why.

This is frankly more complicated than I like -- conditionals based on something outside the path itself are an abhorent complication, and basedirs being "." just plain feels more correct to this author than "./" -- but correct treelike sort orders is trump.

The alternative is to implement path handling and sorting structures that internally convert everything into a series of string slices.  That might be cleaner, but it's also most certainly going to cause a lot of conversion costs on the interface from that to any other path calls (including, right up till the end of the day, everything the kernel and syscalls feel about paths), and also a lot of mallocs.  Someday it would be nice to prototype it and do a real comparison benchmark, but from here my money's on canonicalization to ascii-sortable being the fast route.

Signed-off-by: Eric Myhre <hash@exultant.us>