Add option to test time with truncated precision.
Use this in tar tests.  Uncover a major problem with tar outputs: previously, we were computing hashes with more information than the output stream actually ended up with!  Hashes should always be computed with exactly the same level of information as the data stream supports, so that precise round-trips are possible.  Fixed.

Signed-off-by: Eric Myhre <hash@exultant.us>