IO rework now centers around transport drivers.
And, added back in data system kind names, so we can now talk about examples of dispatcher systems.  There's now a dispatcher system; it manages to have the same interface as single Transmats do.

Prior commits tell the bloody details of the progress of science, but basically the driving force behind this focus is realization that some sort of manager structure for a group of dataset manifestations is actually the normative state of being, because somebody with an algorithm has to be able to own an area, so that they know how to handle it when re-loading from cold.

A lot of dispatcher setup is coming out as imperative programmatic stuff, but that's probably going to work out just fine.

Signed-off-by: Eric Myhre <hash@exultant.us>