Starting draft work on assemblers.
Drafted a test and example of what results should look like; still mostly husks of interfaces with implementations to come.

Placers, like other things before them, are sprouting a return value that allows for their future cleanup.  The temptation to wrap all that out of sight with a function-passing system is strong, but it's key to remember that all these systems are supposed to also be prepared to work on a cleanup situation where the original invokers died unexpectedly.

Signed-off-by: Eric Myhre <hash@exultant.us>