Breaking ground on IO backed by s3 storage.
Going to go about this in a very MVP-oriented way: tarballs streamed in an out, because we know how to preserve file attributes in a widely-understood way by doing this.  (Another existing system that describes a FS<->S3bucket mapping is the s3fs-fuse project; we could also work with something like that in the future.  For now, we don't want to deal with fuse, large c programs of questionable portability, or deal with something that puts latency in the middle of a job and prefer to locate it predictably on the provisioning and teardown edges.)

Adding a dependency on s3gof3r, which appears to be a good, solid readymade API to stream large files to s3.  (In the future this API is something we might actually prefer to use via an exec path, so we can have it behave as an optional tool rather than bloating the main binary for folk who aren't interested in it... but we'll do direct binds for now; it's pretty tiny, anyway.)

Signed-off-by: Eric Myhre <hash@exultant.us>