S3 input and round-trip tests.
Having these common tests is paying huge dividends.  We should have more of them, to be sure, but even this far it's a huge enabler.

There's a *lot* of identical code here between the s3 inputs and outputs.  It's messy.  This will be fixed soon: There's a heavy IO rework floating in progress in another branch which will involve refolding this code to address that issue -- it's so far it's consistently emerged that pairs of inputs and outputs have way more common with their pair than they do with their category.

Signed-off-by: Eric Myhre <hash@exultant.us>