Convert a test to use the new requires system.
It's a little unpretty to have the requirements separated from that method's first appearance, though.  Which is a bummer, because it's enforced by the trailing varargs rule.  I might try to hack around that.  A chaining fluent thing would do the trick, but seems kinda heavy, so maybe just punting to the `items ...interface{}` that's already running rampant in goconvey.

Signed-off-by: Eric Myhre <hash@exultant.us>