Some kinds of transmat take factories; others don't really care.
There's no reason to be uniform if you're not a plugin implementing that entrypoint interface.  It makes sense for the cacher to take command over its contained stuff, and it doesn't make much sense for dispatches to be in charge of construction; so now they just don't have the same interface, and really, nobody cares.

Signed-off-by: Eric Myhre <hash@exultant.us>