Put scan function back on Transmat.
We might punt on having multi-stage output stuff wherein output arenas are set up first and data shovelled into them, and just have scans accept an existing fs as the default interface for now.  It very much seems correct to have assemblers capable of putting together filesystem arrangements where some components do a write back to their src dirs, but...  There's a lot of sullied whiteboard on this, but right now we don't have implementations that need mounts to live, they seem like the exception rather than the rule... the one major payoff is when shuttling output through assemblers means we can get data in the right resources the first time when we have heterogenous local disks, and that's a fairly distant niche need.   So we're just going to take that whole mess off the table for now.

Signed-off-by: Eric Myhre <hash@exultant.us>