Implement Dir as the new Transmat interface.
This is just a wrapper around the previous interface; this is so that we can do a migration from the old code in three phases.  Phase 1 is these wrappers; Phase 2 is cutting over executors to use the new APIs; and Phase 3 is getting the wrappers ironed back out.  Ugliness for the duration is strong (we're converting back and forth between completely useless channels), but it'll be worth it for being able to move incrementally.

Signed-off-by: Eric Myhre <hash@exultant.us>