Treat paths in tar output correctly:
The `Apply` semantic is to take a whole path with no further rehashing.  And yes -- this does make for some appearance of redundancy, since this means the output system is ignoring def.Output.Location.  This is intentional, and the weirdness stems from the fact we're passing user-land config API structures way too deep, and this will all be improved by the next passover of IO refactors.

Also tweak the prefixy thing to save the tar with "./" pattern paths, which makes it act more like our own direct tar writers.

Signed-off-by: Eric Myhre <hash@exultant.us>