Export HashingReader.  Fix dir input to use.
Also quietly fixes a bug in dir inputs: permissions bits on files were not being correctly set, and now are by virtue of reusing the already-tested paths in fs.PlaceFile.  (Tests coming as part of the io-reboot refactor will cover this.)

Yayy for diffs that are net removals of code!

Signed-off-by: Eric Myhre <hash@exultant.us>