Reconnect s3 as transmat.
This required some changes of the existing source to handle null URIs like some of the new tests require.  It's a big and ugly diff; there's no point in making it pretty since it's going to just change again when we upgrade that API to use slices, which will get rid of this null-scheme hack.

Signed-off-by: Eric Myhre <hash@exultant.us>