More progress log messages.  Announce assembly.
Our overall output is still... extremely conmingled.  You can tell we haven't seriously looked at getting these to come out in an order even remotely close to the order things are actually... going on.  Which, indeed, wasn't an objective.  But now that I'm looking at this, I kinda want that to be... possible, at least.

Signed-off-by: Eric Myhre <hash@exultant.us>