Affix base dir attributes on tar inputs.
This is a fix for a bug that would let non-normalized data slip in from tars that lacked attributes for the base dir.

This has usually been masked in practice by dirs that happened to be created with identical attributes to our defaults for hashing, but we should really do it right, here.

Signed-off-by: Eric Myhre <hash@exultant.us>