FIX MODES, resulting from adding tests for bouncing a foreign tar stream.
THIS IS A BREAKING CHANGE and a giant fucking flag day on hashes.  Here it is, your "we're pre-alpha" admonition in practice!

Whitelist the first 07777 bits of mode when feeding the hasher.  These are the only bits we pass out in our chmod/open syscalls; and we absolutely should be using the same filter in both scenarios.

Turns out golang's tar attribute reader puts all sorts of bits in the higher parts of the mode which are spec-compliant but totally redundant.  GNU tar (I'm staring at 1.27.1, if relevant) does *not* put these bits in the headers in its output.

This fixes the possibility of two tarstreams that are semantically the same to us being treated as different hashes; as such, despite being a startling flag day, this does not constitute a security issue (the whitelisting on syscalls was actually *more* strict and consistent than the one on the hashes; security/integrity issues only arise if it goes the other way, where two different data can make it to syscalls but have the same hash).

Signed-off-by: Eric Myhre <hash@exultant.us>