Systematic review of IO system errors.
We've learned a lot about the practical facts of where errors come from in IO systems and what kind of responses are possible, so I'm going to reorganize around that as I'm moving code from old locations into the new transmat interfaces.

This will be expected to completely replace both input/errors.go and output/errors.go by the end of the conversion march.

Signed-off-by: Eric Myhre <hash@exultant.us>