Add URI parsing to tars.  Support http!
Parsing storage locations as URIs gives us a functioning ability to swap out disparate warehouse locations, as long as they still using the same systems to map from active filesystem to cold storage form.  First up is making tars act the same way regardless of if they're from a local filesystem or a remote http host.

This means:

- We can use HTTP storage
- Since the HTTP storage is anonymous, we can now offer that for anyone to use to do bootstrap
- Both HTTP and local file tar usage can now use content-addressible mode, which means using a local folder for bulk job output gathering is now way (way) easier.

This is *not* necessarily going to be a perfect, end-all solution to the matrix of transports and storage locations versus storage formats and plugins.  Specifically, it does *not* propose any new ways to handle things like secret management for authenticated systems, or afford much extensibility (query params, sure, but when we're multiplexing onto http/rest patterns already?  Not a solution; not isolated.).  Additionally, we're overloading the "scheme" field in order to set the "is content-addressible?" bit on systems where that is optional.

Future work will likely include more complex approaches that support more structural data.  We'll see exactly how that bears out: the API should also aim to be simple first and support complexity as necessary, so we'll probably try to navigate a smooth upgrade path where one-liner dead simple URIs still work for basic cases.  (Thinking out loud: maybe "myS3ConfAlpha://bucket/path" and then have a separate section of task config later that specs out what exactly that's supposed to drive and what creds it should have?  Something like that would let us specify credentials once for a repeatedly used warehouse, which seems important.)

Side note: Code is increasingly referring to "warehouse coordinates", because that seems to better capture the idea that a warehouse can contain lots of different stuff, and we're only talking about a shipping and distribution detail.  This isn't uniform but we'll probably continue to do a slow-rolling refactor in that direction.

Some new test coverage is lagging behind.  Turns out we need to reason about tempdir behaviors pretty much all the way at the top.  Will require a bit of a refactor on the common tests.

Signed-off-by: Eric Myhre <hash@exultant.us>