Start standardizing executor tests.
I'm not going to fight with nsinit yet, but we can at least migrate all this code to hit the interface instead of literal chroot implementation structs.

Some cleanups and reorgs will come along the way.

Requirements and tempdir setup will be kept at the top, in keeping with what we learned over time from working on the transmat shared tests.

Signed-off-by: Eric Myhre <hash@exultant.us>