Define no-such-cwd to be a TaskExecError.
Letting it be reported as a NoSuchCommandError would be a ridiculous leak of implementation details of this one executor, and frankly, doesn't make much sense even in isolation.  So, do a check before invoking the part of exec that gives the unhelpful error.

This may make sense to report as a special/specific NoSuchCwdError so that it can be reported as a config error, but that's a decision I'll defer for later work.

Signed-off-by: Eric Myhre <hash@exultant.us>