Update readme; reference the demo script.
The tests are good to know about too, but the demo is more built to be looked at.  And we can ship the demo to work with pre-made repeatr binaries (which we should totally tackle sometime soon; adding "go install golang!" to the readme made me suddenly realize that's really betraying the whole look-ma-it-ships claim to fame).

Signed-off-by: Eric Myhre <hash@exultant.us>