Set tmpdir perms in build script.
It's easier to do this out here than leave it to the code inside repeatr.  As a standalone executable, repeatr doesn't initate the base dir to be sticky and everyone-writable; in a testing context, we want exactly that to already have happened, so there isn't a messy result when someone runs the tests with root and then again later without priviledges.

Also, remove references to "1755" because... that probably doesn't mean what you might have thought it means when seen as golang's os.FileMode.

Signed-off-by: Eric Myhre <hash@exultant.us>