Hm, maybe printing errors to dev null wasn't wise.
Especially since in addition to a test that didn't actually do anything except print errors to io.Discard, we also have a CI script that assumes these commands are going to exit non-zero if they're broken, and *they don't*, they just toss out a big mumble of errors and then exit as if nothing went wrong.

So they've -- perhaps not surprisingly -- been broken for weeks.  On green tests.

I regret allowing this through code review in the first place.

The new behavior to `repeatr run` is to exit non-zero if any jobs failed.  This fixes the immediate problem of non-apparent failures.  There continue to be no tests for actual multiple jobs in one invocation.  That feature was frankly never spec'd out; it may be removed or replaced.

Changes in this commit include fixes to each of the out-of-date configs, now that their failures would actually (correctly) break CI.

Signed-off-by: Eric Myhre <hash@exultant.us>