Emit config parse errors as CLI error (no stacks).
Nobody needs to see stack traces from inside the json parsing library when their config file has a typo.

Note: There's a ConfigError type, and we're not using it here.  Why?  Because that type is in the io/integrity package.  It makes sense in that package.  That package is supposed to be severable and usable as a library, so we're not moving it up into CLI or something broad like that; and it doesn't really make sense for the CLI to emit an integrity.Error, or rewrap that.  There's also no obvious upside to making subtypes of CLIError since that's basically just an indicator that "this is already formed to be human readable".  Thus: this.

Signed-off-by: Eric Myhre <hash@exultant.us>