Correctly check for nil errors.
Interface conversion makes this trickier than it should be -- the way the struct field `result.Error` is used in exec_tests.go:97 means we can hit the http://golang.org/doc/faq#nil_error problem.

What really makes it a trainwreck is that it's capable of generating spooky action at a distance: the panic is from some code in the spacemonkey library getting a nil pointer panic *right after it checks that exact value being nil*, and it's due to a way I'm satisfying an interface used internally to goconvey.  :(  This does not seem like the best considered of design choices in golang.  It's documented, sure, but the complexity it introduces is just bonkers -- of all the reasons I'd expect to need to import the reflect package, checking if a parameter to my function is *actually* nil is certainly not one of them.

Signed-off-by: Eric Myhre <hash@exultant.us>