Add input stream parameter to Executors.
This is nil for all existing uses.  Adding a special, interactive-mode command for single jobs will be able to feed in an input stream, though -- at the cost of effectively discarding repeatability, of course (so it will be a whole different command, *not* a flag on `run`).

We don't currently feed the stdin stream into the streamer mux for storage.  This might be up for rehash later, but seems like the right thing to do -- shell echo is a thing, normally; if you turned that off, one presumes you had a reason.

Signed-off-by: Eric Myhre <hash@exultant.us>