Begin addition of a playground mode.
`repeatr twerk` is for... let's call it trial working: a playground where you can set up a whole environment as normal, but we also give you stdin and a shell.  If you don't specify a rootfs, etc, you'll just get one.  Repeatability is shot before you get started... but this can still be a handy pressure release valve for debugging, playing with outputs, experimenting before writing a formula, etc.

This'll need lots more stuff before it's useful: right now it's hardcoded ubuntu and bash.  We'll want overrides, we'll want the ability to fork a cwd (I'm writing this to replace a shady shell script that straps together docker and a COW cwd, frankly, because for all its jank that is just a crazy useful thing), we'll *badly* want ttys to "do the right thing", etc.  :soon:

Signed-off-by: Eric Myhre <hash@exultant.us>