First pass on a git transmat.
This is restricted to read-only usage (we may support scan/upload some day, but what's a helpful way to engage with history concepts is not yet decided), and there are many other caveats (git isn't as precise as our other transports), and it's not yet well isolated, but it's essentially functional.

This makes it possible for a project to track its own source as an input easily, which should be very exciting.

Signed-off-by: Eric Myhre <hash@exultant.us>