Remove requirement of git in generate step.
By specifying the "GITCOMMIT" variable in advance, you may ignore attempts to use the git command.  This is both convenient if you're doing a build in an environment without git installed, or if you're interacting with a source export that doesn't include .git dirs (which, you'll notice is how the git transmat will work by default).

Signed-off-by: Eric Myhre <hash@exultant.us>