Define scanning a nonexistent path to return nil.
Retrofit all existing transmats to behave in this way and added to set of standard specs.

Previously most transmat implementations would panic (worse, with a very nondescript error).  Given that output dirs not having been created is a very common side effect of ~something~ going wrong in a contained script, it seems clear we should be relatively calm about it.

Also considered: requiring a zero content hash be a well defined operation for all transmats.  But that doesn't really seem any more useful.

Signed-off-by: Eric Myhre <hash@exultant.us>