Start hammering logging through provisioning steps
It'll be nice to have a little color through there, at least.  This may not be the ultimate state of this code, either, though -- this is an area where interactive use might still benefit from a special touch.  But, we don't have fully baked concepts of progress bars for most of this either, so, one step at a time.

Configuration is currently nonexistent.  I'm going to assume this is heading for human eyes.  There's a good suggestion of how to do configurability (at a library level, at least, which is also not exactly the domain problem here...) at https://github.com/inconshreveable/log15/issues/61#issuecomment-122550114 which we may consider adopting as it becomes an issue.

Some sprintfs are used for the messages.  This is on the theory that those lines should be readable and meaningful without scanning down the whole context list (maybe I'd feel differently if context was before, then message, then parameters from the log call?)... this may not be a scalable or reasonable idea; it just looked weird to say "starting materialize" and *nothing else*.  So, going with it for now; may change.

Signed-off-by: Eric Myhre <hash@exultant.us>