Thinking out loud.
I already started removing the multi-formula feature once, then talked myself back into it (surely it would actually make sense to pipe a bunch of runs into a big heap of records and just sit back and walk it later?).  But it really is crap for almost every kind of composition in the shell.

Signed-off-by: Eric Myhre <hash@exultant.us>