Remove underdefined multiple formula support from run command.
As discussed over the last few commits, this is just too much trouble and is blocking the way of more immediately useful semantics and scripting.  We'll re-introduce a more parallel mode later; right now, you can frankly do as much in bash with `&` (no, really, the scheduler implementations are such placeholders you're almost better off; and we've been explicitly supporting acting like a process without daemoning/locking/centralizing this whole time for a reason).  When re-introduced, multi-formula mode will have a much better doc on what the heck one is supposed to do to actually watch jobs; this will require more of an API story in general.

Signed-off-by: Eric Myhre <hash@exultant.us>