Repeatr, repeat thyself!
Demonstration of repeatr building repeatr repeatedly.  This is pretty exciting.

Now, mind, we don't actually have deterministic output yet (but we're closing in very, very rapidly! :D).  Here's a list of things that need to be improved:
  - We need to enable filters!
  - You see what I did there with git "HEAD"?  Heh, yeah, that should be illegal.
  - Json sucks as a config format if you need to embed things.  We either need to start working with different formats, or make tools to help.
  - Optionally, improve repeatr test asset bootstrap.  (We're not currently doing tests in this formula because they rely on bootstrap paths that rely on shell scripts that rely on... wget.  Which is somehow not in our default ubuntu image.  There's no compelling reason for these to be legacy mode anymore.)

That's... actually, yeah, not much of a list.  That first item is big enough to be its own feature branch still, but the rest are just minor friction with approachable workarounds (we'll still focus fire on the usability issues there eventually, but they're not blockers for coup-de-grace on repeat-thyself).

Notably absent from that todo list: fixing the golang compiler.  Yeah, turns out, golang is actually deterministic out of the box with its outputs.  I'm pretty wonderfully thrilled about that.

Signed-off-by: Eric Myhre <hash@exultant.us>