Remove jq dependency from repeat-thyself.sh.
The replacement is a hilarious series of sed/tr/grep hacks (yes, all of them, because fuck you, that's why; they're all equally a mistake, so in for a penny in for a pound) which does not actually escape json (obviously) but comes close enough on this example.

jq is amazing and wonderful and even remarkably well done work on the portability front, but it's not desirable for repeatr workflows to encourage system dependencies.  (Sed's invalid by that reasoning too, but A: slightly more people have it out-of-the-box and B: it makes me feel bad, so it's more likely to get fixed.)

Signed-off-by: Eric Myhre <hash@exultant.us>