List public url in repeat-thyself for travis.
Oh my: travis uses shallow clones, and also evidentally (as far as I can gather) a version of git that doesn't know how to clone from a shallow clone (I'm on 1.9.1 locally, which has no problem with that).

Fodder for the argument that the git plugin needs isolation more desperately than any other plugin to date.  And yet, I don't want to require chroot and thus root for minimum viable git usage.  So, at some point we're going to have to go down an adventuresome road of static linking and fun repackaging for git if we want this licked.

Signed-off-by: Eric Myhre <hash@exultant.us>