Detect DataDNE in git and raise appropriate error.
Many other failures are still lumped in a catchall.  Previous integrations I've done with git leave me feeling quite confidient that pretending to have full coverage of the possible failure modes of git porecelain commands is tantamount to insanity, and we're just going to admit that there will always be a catchall.  We might chip away at it (like this, with the DataDNE), but it will never leave.

First draft of this added the command output as an errors.DataKey, but, frankly, the value added on those is... almost negative in this situation, because they end up not being printed by either the errors.Message function or the toString behavior.  Rant rant formatting.  I dream of replacing spacemonkey errors at this point, with something very similar and yet not at all alike.

Signed-off-by: Eric Myhre <hash@exultant.us>