Cache git checkouts.
This is super, super primitive: invocation of the same repo with different hashes will still not result in clone or object reuse, either in download savings nor in disk savings hardlinks.  It will, however, speed up repeated uses of the same commit hash.

Additionally, this does not separate hashes per repo URL.  This... is probably reaosnably safe, in practice, for now.  SHA-1 is a little shakey, though, and I'd like to get to higher ground.  The question on how to do that is all in git's hands, however, unless we can think of some super clever ways to add layers of hash (and baring some extremely clever future ideas, the usability story there would probably be pretty unappealing).

Signed-off-by: Eric Myhre <hash@exultant.us>